Update domain command to send summarized images of domain schedules
Previous implementation of this command sends a lengthy embedded message that contains the details of the specified day's domain schedules. This is replaced in favor of sending a link to an imgur image that contains visual summaries of the domain schedules.

Also added a `tomorrow` option as argument for the command.